% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{make_df_list}
\alias{make_df_list}
\title{Make a named list of distribution functions}
\usage{
make_df_list()
}
\value{
Named list of (density, quantile, generator) triples for distributions
}
\description{
Constructs the default nested list of distribution functions used to animated
normal QQ plots. By default, the list contains 
  Normal  Student t_3  Student t_6  Gamma(3)  Beta(0.6,0.6)   Exotic
Each member of the list has 3 named elements [
   d = density function  (eg, dnorm)
   q = quantile function (eg, qnorm)
   r = random generator  (eg, rnorm) ]
Each of these functions must take a single argument; others must be bound externally.
For examples, see the function make.normal.dist or make.t.dist.
}
\examples{
make_df_list()

}

